+++ src/main/java/roomescape/reservation/infrastructure/vo/ThemeBookingCount.java
@@ -1,6 +1,5 @@
 package roomescape.reservation.infrastructure.vo;
 
-import roomescape.reservation.domain.BookedCount;
 import roomescape.theme.domain.Theme;
 
 public record ThemeBookingCount(Theme theme, BookedCount bookedCount) {
